+++ src/hooks/useVirtualList/index.ts
@@ -39,6 +39,7 @@ export default <T = any>(
       [0],
     );
   }, [heightList]);
+
   const marginTop = useMemo(() => {
     return rowTopOffsetList[rowIndex.start];
   }, [rowTopOffsetList, rowIndex.start]);